fix(core): update name validation for handler v4

+++ libvctrl_core/src/validate/name.rs
@@ -4,7 +4,7 @@ pub fn validate_name(name: &str) -> Result<(), VctrlError> {
     if name.is_empty() {
         return Err(VctrlError::InvalidName("name is empty".into()));
     }
-    if name.len() > MAX_NAME_LENGTH {
+    if name.len() > MAX_NAME_LENGTH as usize {
         return Err(VctrlError::InvalidName(format!(
             "name exceeds maximum length {MAX_NAME_LENGTH}: '{name}'"
         )));
